1.修改sim卡注册网络时，添加CREG:0,1本地网络,CREG:0,5漫游网络，余情况为异常情况。2.修改显示历史数据会出现乱码问题，原因是fileSeek的变量超出了范围。

+++ Src/fatfs.c
@@ -309,7 +309,7 @@ ErrorStatus FATFS_FileSeekBackwardOnePack(void)
 /*******************************************************************************
  *
  */
-ErrorStatus FATFS_FileSeek(WORD byte)
+ErrorStatus FATFS_FileSeek(DWORD byte)
 {
 	if (byte <= objFile.fsize)
 	{
